Small Town Complete
There are two building and a really big one that is supposed to act like a hotel/apartment of sorts.
There are two stalls too! In the off chance we want to see some minor stuff in the first town.
AND THERES MUSIC o(^▽^)o

Have good night ya'll. Tell me if there is anything that needs to be updated/changed!